requirements.txt
# Google ADK Agent Starter Kit - Dependencies

# Core Google ADK dependencies
google-adk>=0.2.0
google-generativeai>=0.3.0
google-cloud-aiplatform>=1.36.0

# Authentication and API access
google-auth>=2.22.0
google-cloud-core>=2.3.2
google-api-python-client>=2.100.0

# Web server for local development
fastapi>=0.103.1
uvicorn>=0.23.2
websockets>=11.0.3

# Environment variable management
python-dotenv>=1.0.0

# Utilities
pydantic>=2.4.2
typing-extensions>=4.7.1
certifi>=2023.7.22
requests>=2.31.0

# Testing
pytest>=7.4.2
pytest-asyncio>=0.21.1
pytest-cov>=4.1.0

# Documentation
sphinx>=7.2.6
sphinx-rtd-theme>=1.3.0

# Development tools
black>=23.9.1
isort>=5.12.0
flake8>=6.1.0
mypy>=1.5.1
